fix: bump PutarFlix to v5 and fix timeout type mismatch
- Bumped build.gradle.kts version to 5 for the compile fix
- Changed request timeout arguments from Int conversion to Long values
- Fixed app.get and app.post timeout parameter mismatches in PutarFlixExtractor
- Preserved the v4 playback resolver logic without changing provider behavior

+++ PutarFlix/src/main/kotlin/com/putarflix/PutarFlixExtractor.kt
@@ -465,7 +465,7 @@ internal object PutarFlixExtractor {
                 app.get(
                     url = url,
                     referer = referer,
-                    timeout = REQUEST_TIMEOUT_MS.toInt()
+                    timeout = REQUEST_TIMEOUT_MS
                 ).document
             }.getOrNull()
         }
@@ -477,7 +477,7 @@ internal object PutarFlixExtractor {
                 app.post(
                     url = url,
                     referer = referer,
-                    timeout = REQUEST_TIMEOUT_MS.toInt(),
+                    timeout = REQUEST_TIMEOUT_MS,
                     headers = mapOf(
                         "X-Requested-With" to "XMLHttpRequest",
                         "Content-Type" to "application/x-www-form-urlencoded; charset=UTF-8",
